requirements.txt
flask
azure-identity
azure-storage-blob
gunicorn
pandas
openpyxl
python-dotenv
azure-identity
azure-keyvault-secrets
